Move session mock-up to another file
